Updated gradle wrapper version

+++ src/main/java/com/google/gson/internal/LazilyParsedNumber.java
@@ -50,7 +50,7 @@ public final class LazilyParsedNumber extends Number {
     try {
       return Long.parseLong(value);
     } catch (NumberFormatException e) {
-      return new BigInteger(value).longValue();
+      return new BigDecimal(value).longValue();
     }
   }
